+++ src/app/shared/contact-us/contact-us.component.ts
@@ -1,4 +1,9 @@
-import { Component, OnInit, Inject } from '@angular/core';
+import { ConfigOptionsService } from './../../core/config-options.service';
+import {
+  GeneratorService,
+  TokenGenerator
+} from './../../core/generator.service';
+import { Component, OnInit, Inject, Optional } from '@angular/core';
 import { ConstantsService } from './../../core/constants.service';
 
 @Component({
@@ -7,9 +12,24 @@ import { ConstantsService } from './../../core/constants.service';
   styleUrls: ['./contact-us.component.scss']
 })
 export class ContactUsComponent implements OnInit {
-  constructor(@Inject(ConstantsService) private constant: any) {
-    console.log(this.constant);
-  }
+  tokens: string[] = [];
 
-  ngOnInit() {}
+  constructor(
+    @Inject(ConstantsService) public constant: any,
+    @Inject(GeneratorService) private generatorService: TokenGenerator,
+    @Optional() private configOptionsService: ConfigOptionsService
+  ) {}
+
+  ngOnInit() {
+    let i = 10;
+
+    while (i--) {
+      const token = this.generatorService.generateToken();
+      this.tokens.push(token);
+    }
+
+    if (this.configOptionsService) {
+      this.configOptionsService.saveConfig(this.constant);
+    }
+  }
 }