+++ src/app/shared/shared.module.ts
@@ -4,17 +4,20 @@ import { CommonModule } from '@angular/common';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { FormsModule } from '@angular/forms';
 import { HoveredDirective } from './hovered.directive';
+import { ClickNMarkDirective } from './click-n-mark.directive';
 import { ContactUsComponent } from './contact-us/contact-us.component';
 
 @NgModule({
-  declarations: [HoveredDirective, ContactUsComponent],
+  declarations: [HoveredDirective, ClickNMarkDirective, ContactUsComponent],
   imports: [CommonModule, MaterialModule, FlexLayoutModule, FormsModule],
   exports: [
     HoveredDirective,
     CommonModule,
     MaterialModule,
     FlexLayoutModule,
-    FormsModule
+    FormsModule,
+    ContactUsComponent,
+    ClickNMarkDirective
   ]
 })
 export class SharedModule {}