+++ changelog.md
@@ -16,3 +16,14 @@
 4. move cart into popup
 5. implement quantity change (while rotating mouse wheel ), delete product and price recalculation functionality
 6. created directive AppHovered to change background of hovered elements.
+
+
+## Task 3
+
+1. Add cleanCart method to the CartService. Other methods we implemented
+2. Add LocalStorageService
+3. Add ConfigOptionsService
+4. Add ConstantsService
+5. Add GeneratorService and register it as Factory
+6. Add app-contact-us component
+7. Add click-n-mark.directive, apply it to the list of random strings